iox-#1560 Move 'BestFittingType' to the algorithm header
