+++ iceoryx_posh/include/iceoryx_posh/internal/popo/building_blocks/condition_listener.hpp
@@ -29,7 +29,7 @@ namespace popo
 class ConditionListener
 {
   public:
-    using NotificationVector_t = cxx::vector<cxx::BestFittingType_t<MAX_NUMBER_OF_NOTIFIERS>, MAX_NUMBER_OF_NOTIFIERS>;
+    using NotificationVector_t = cxx::vector<algorithm::BestFittingType_t<MAX_NUMBER_OF_NOTIFIERS>, MAX_NUMBER_OF_NOTIFIERS>;
 
     explicit ConditionListener(ConditionVariableData& condVarData) noexcept;
     ~ConditionListener() noexcept = default;
